Turn template into file Move all helpers into helpers.rb
