+++ metadata.rb
@@ -7,7 +7,7 @@ license 'Apache-2.0'
 description 'Configures sysctl parameters'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
 version '0.10.2'
-chef_version '>= 13.0' if respond_to?(:chef_version)
+chef_version '>= 12.5' if respond_to?(:chef_version)
 
 supports 'amazon'
 supports 'ubuntu', '>= 14.04'
