+++ test/cookbooks/test/metadata.rb
@@ -1,7 +1,3 @@
 name 'test'
-maintainer 'Viverae, Inc.'
-maintainer_email 'cookbooks@viveraehealth.com'
-license 'Apache-2.0'
-description 'Wrapper cookbook to test sysctl resources'
 version '0.1.0'
 depends 'sysctl'
