+++ test/cookbooks/test/recipes/default.rb
@@ -1,22 +1,3 @@
-#
-# Cookbook Name:: test_sysctl
-#
-# Copyright 2013-2014, OneHealth Solutions, Inc.
-# Copyright 2014, Viverae, Inc.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-#
-
 sysctl_param 'vm.swappiness' do
   value 19
 end
@@ -26,5 +7,6 @@ sysctl_param 'kernel.msgmax' do
 end
 
 sysctl_param 'kernel.msgmax' do
+  value 9000
   action :remove
 end
